+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
-v0.82.1
+v0.84.0
 
-- keytrans: Unify errors with other typed APIs
+- keytrans: Verify signatures from all auditors
 
-- Locally-closed WebSocket connections now explicitly send close code 1000 instead of an empty close frame
+- Java: Fixed `IdentityKeyPair(byte[])` to correctly declare that it throws InvalidKeyException.
+
+- Node: Add BackupsJsonExporter, to convert Backup proto objects to human-readable JSON for export.